requirements.txt
click==6.7
colorama==0.3.7
contextlib2==0.5.4
coverage==4.2
db==0.1.1
Django==1.9.6
djangorestframework==3.3.3
docker-py==1.7.2
dockerpty==0.4.1
docopt==0.6.2
docutils==0.12
dominate==2.3.1
ecdsa==0.13
futures==3.0.5
get==0.0.20
gunicorn==19.6.0
httpie==0.9.9
idna==2.6
itsdangerous==0.24
jdcal==1.2
Jinja2==2.9.6
jmespath==0.9.0
Mako==1.0.4
MarkupSafe==1.0
models==0.9.3
mysqlclient==1.3.9
numpy==1.11.1
olefile==0.44
packaging==16.8
pathspec==0.3.4
post==0.0.12
public==0.0.37
py==1.4.31
pyasn1==0.1.9
pycparser==2.14


